requirements.txt
apache-airflow==2.7.1
dbt-core==1.6.0
dbt-bigquery==1.6.0
google-cloud-bigquery==3.11.4
pyspark==3.4.1
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.2
jupyter==1.0.0
black==23.7.0
flake8==6.1.0
isort==5.12.0 